backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
tensorflow==2.15.0
Pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==21.2.0
scikit-learn==1.4.0
matplotlib==3.8.2
tqdm==4.66.2